refactor(sidecar): relocate SocketPathResolver into Reli\Sidecar\Client
The sidecar client (Reli\Sidecar\Client) is intentionally FFI-free so user
applications can depend on it without ext-ffi. SocketPathResolver — the only
non-Client class the client imported — lived under Reli\Inspector\Sidecar,
which couples the client to the inspector tree even though the resolver
itself uses nothing beyond the standard library.

Move the class (and its test) into Reli\Sidecar\Client so the Client
namespace is self-contained. This is a prerequisite for splitting the
client out as a standalone, FFI-free package via subtree split.

No behavior change — pure namespace relocation.
