+++ tests/Sidecar/Client/SocketPathResolverTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Reli\Inspector\Sidecar;
+namespace Reli\Sidecar\Client;
 
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
